feat: User 인증 및 missionId 타입 수정

+++ backend/src/main/java/com/example/onboarding/mission/dto/MissionResponseDto.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 @Builder
 public class MissionResponseDto {
 
-    private Long missionId;
+    private int missionId;
     private String title;
     private String description;
     private LocalDateTime deadline;
